Create resources to allow local clusters to pull from ecr (#6)

+++ knitter/bootstrap/local/destroy_zLifecycle.sh
@@ -8,23 +8,30 @@
 # proprietary to CompuZest, Inc. and are protected by trade secret or copyright
 # law. Dissemination of this information or reproduction of this material is
 # strictly forbidden unless prior written permission is obtained from CompuZest, Inc.
+LOCATION=$1
 
-kubectl port-forward service/argo-cd-argocd-server 8080:80 -n argocd
+if [[ -z "$LOCATION" ]]
+then
+    echo "Error: Please pass the name of the environment you'd like to destroy to this script"
+    exit 1
+fi
+
+kubectl port-forward service/argo-cd-argocd-server 8080:80 -n argocd &
 
 kubectl get applications -n argocd -o jsonpath='{range .items[*]}{.metadata.name}{"\n"}{end}' | xargs kubectl patch applications  -p '{"metadata":{"finalizers":[]}}' --type=merge -n argocd
 kubectl patch crd applications.argoproj.io -p '{"metadata":{"finalizers":[]}}' --type=merge -n argocd
 argocd app delete $(argocd app list -o name)
 
 argocd cluster rm sandbox
-argocd repo rm git@github.com:CompuZest/compuzest-dev-a-zlifecycle-il.git
+argocd repo rm git@github.com:CompuZest/compuzest-$LOCATION-lzlifecycle-il.git
 argocd repo rm git@github.com:CompuZest/helm-charts.git
 
 cd ../../../zlifecycle-provisioner/k8s-addons
 terraform init
-terraform workspace select 0-local
+terraform workspace select 0-$LOCATION
 terraform init
 
-terraform destroy -auto-approve -var-file tfvars/local.tfvars
+terraform destroy -auto-approve -var-file tfvars/$LOCATION.tfvars
 
 echo ""
 echo ""
@@ -35,6 +42,6 @@ echo "-------------------------------------"
 
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-    k3d cluster delete sandbox-k3d
+    k3d cluster delete $LOCATION-k3d
 fi
